MANTA-902: Monitor DNS for record changes and use connection pools
This is a big change in terms of semantics, as the client, when given
a DNS name will create a ring of TCP connections to each IP in the DNS
record (RR).  Each operation will then be round-robined across both
remote hosts and local connections, ensuring that any given agent
talking to moray is driving load equally across all available moray
CPUs.

In addition, this client will periodically poll DNS looking for
"dropped" or added servers, and adjust the pools as appropriate.  Note
that right now this is highly invasive, as the client just blatantly
closes all old connections and replaces with new ones, as it's assumed
this is an infrequent event.